+++ spec/vizcore/dsl/engine_spec.rb
@@ -17,6 +17,7 @@ RSpec.describe Vizcore::DSL::Engine do
             shader :gradient_pulse
             map frequency_band(:low) => :intensity
             map beat? => :flash
+            map beat_confidence => :sync_strength
             map beat_pulse => :wobble
           end
         end
@@ -38,6 +39,7 @@ RSpec.describe Vizcore::DSL::Engine do
       expect(layer[:mappings]).to include(
         { source: { kind: :frequency_band, band: :low }, target: :intensity },
         { source: { kind: :beat }, target: :flash },
+        { source: { kind: :beat_confidence }, target: :sync_strength },
         { source: { kind: :beat_pulse }, target: :wobble }
       )
     end